requirements.txt
Django==1.9.6
djangorestframework==3.3.3
oauthlib==1.1.1
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.19
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
